automatically backed up by learn

+++ index.js
@@ -21,5 +21,5 @@ function dec(n) {
   return (n -= 1)
 }
 function makeInt(n) {
-  return (parseFloat('n'))
+  return (parseInt('n'))
 }
